split into file
